+++ js/app/controllers/maincontroller.js
@@ -343,6 +343,22 @@ function ($rootScope, $scope, $timeout, $window, $document, ArtistFactory,
 		$scope.collapseNavigationPaneOnMobile();
 	};
 
+	// Compact/normal layout of the Albums view
+	$scope.albumsCompactLayout = (Cookies.get('oc_music_albums_compact') === 'true');
+	$scope.toggleAlbumsCompactLayout = function(useCompact /*optional, invert current value if omitted */) {
+		if (typeof useCompact === 'undefined') {
+			useCompact = !$scope.albumsCompactLayout;
+		}
+		$scope.albumsCompactLayout = useCompact;
+		$('#albums').toggleClass('compact', useCompact);
+		$rootScope.$emit('albumsLayoutChanged');
+
+		Cookies.set('oc_music_albums_compact', useCompact.toString(), { expires: 3650 });
+
+		// also navigate to the Albums view if not already open
+		$scope.navigateTo('#');
+	};
+
 	$scope.collapseNavigationPaneOnMobile = function() {
 		if ($('body').hasClass('snapjs-left')) {
 			$('#app-navigation-toggle').click();